Implement forking feature as clone of copying
For now the forking feature is an (almost) exact clone of the copy
feature. Differences will arise once it's clearer what both of these
functions should do (and once the changes are implemented for copying).
Once these changes are done for both features, it probably also makes
sense to put common behavior in super classes and similar (there's
already a TODO for that in the code).