Fix error message on signup
The error message appeared because `refreshProfile` was called when the
loggedIn observable was triggered, which was caused because logout was
called even though the user was already logged out on registration.
Now logout is still called, but with a parameter to ensure that the
observable isn't triggered. Also, `refreshProfile` is only called if the
user is actually loggedIn when a change is detected.