Prevent accidental navigation to token validation
If the user now access the login page with the ?validateToken query
parameter and the backend returns a 401 error, the user is redirected to
the same page without the query parameter and now error is shown. This
is because it can happen quite easily that the user navigates to the
address containing the query parameter accidentally.
In all other cases, e.g. when there were too many retries, the error
message is still shown (and the user is not redirected).